pyproject.toml
[tool.poetry]
name = "django-demo-ninja"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
django-extensions = "^3.2.3"
django-ninja = "^1.1.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
django-cors-headers = "^4.3.1"
django-import-export = "^3.3.7"
django-admin-interface = "^0.28.6"
docutils = "^0.20.1"
black = "^24.2.0"
pydantic = {extras = ["email"], version = "^2.6.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
